hw/arm: introduce xenpv machine
Create a new machine xenpv which creates a IOREQ server to connect
with Xen. It also creates a tpm-tis-device which connects to swtpm to
support TPM functionalities.

Xen IOREQ connection expect the TARGET_PAGE_SIZE to 4096, and the xenpv
machine on ARM will have no CPU definitions. We need to define
TARGET_PAGE_SIZE appropriately ourselves.

Signed-off-by: Vikram Garhwal <vikram.garhwal@amd.com>
Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@xilinx.com>

%% original patch: 0010-hw-arm-introduce-xenpv-machine.patch