xen_pvdev: Do not assume Dom0 when changing directory permissions
Instead of forcing the owner to domid 0, read the directory permissions
and add xen_domid with the given permission to it.

Note that for other than Dom0 domain (non toolstack domain) the "driver_domain"
property should be set in domain config file for the toolstack to create
required directories in advance.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>

%% original patch: 0015-xen_pvdev-Do-not-assume-Dom0-when-changing-directory.patch